Moved local inference to separate file.
