+++ external-declarations/tests/source/local-inference/scratch.ts
@@ -1,0 +1,5 @@
+export const x = [
+    "A" as const,
+    "",
+    "B"  as const
+]
